refactor: replace omit to removeWorkspacePackages
