+++ packages/create-codes/src/helpers/remove-workspace-packages.ts
@@ -1,0 +1,7 @@
+export function removeWorkspacePackages(deps: Record<string, string>) {
+  const filteredEntries = Object.entries(deps).filter(
+    ([, value]) => !value.startsWith("workspace:")
+  );
+
+  return Object.fromEntries(filteredEntries);
+}
